requirements.txt
pyyaml==6.0.2
numpy==1.26.4

